[SW-2086] Move classes from org.apache.spark.h2o.utils to ai.h2o.sparkling
