+++ core/src/main/scala/ai/h2o/sparkling/backend/H2OAwareBaseRDD.scala
@@ -17,7 +17,6 @@
 package ai.h2o.sparkling.backend
 
 import org.apache.spark.h2o.H2OContext
-import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.rdd.RDD
 import org.apache.spark.{OneToOneDependency, Partition, SparkContext}
