+++ core/src/main/scala/ai/h2o/sparkling/backend/H2OFrameRelation.scala
@@ -17,9 +17,9 @@
 
 package ai.h2o.sparkling.backend
 
+import ai.h2o.sparkling.backend.utils.ReflectionUtils
 import ai.h2o.sparkling.frame.{H2OColumn, H2OColumnType, H2OFrame}
 import org.apache.spark.h2o.H2OContext
-import org.apache.spark.h2o.utils.ReflectionUtils
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.sources.{BaseRelation, PrunedScan, TableScan}
 import org.apache.spark.sql.types.{Metadata, MetadataBuilder, StructField, StructType}
