+++ core/src/main/scala/ai/h2o/sparkling/backend/Reader.scala
@@ -20,13 +20,12 @@ package ai.h2o.sparkling.backend
 import java.util.TimeZone
 
 import ai.h2o.sparkling.backend.converters.TimeZoneConversions
+import ai.h2o.sparkling.backend.utils.SupportedTypes
+import ai.h2o.sparkling.backend.utils.SupportedTypes._
 import ai.h2o.sparkling.extensions.serde.ChunkAutoBufferReader
 import ai.h2o.sparkling.frame.H2OChunk
 import org.apache.spark.h2o.H2OConf
-import org.apache.spark.h2o.utils.SupportedTypes._
-import org.apache.spark.h2o.utils.{NodeDesc, SupportedTypes}
 import org.apache.spark.unsafe.types.UTF8String
-
 
 private[backend] class Reader(keyName: String,
                               chunkIdx: Int,
