+++ core/src/main/scala/ai/h2o/sparkling/backend/external/ExternalH2OBackend.scala
@@ -22,12 +22,11 @@ import java.util.Properties
 
 import ai.h2o.sparkling.backend.exceptions.{H2OClusterNotReachableException, RestApiException}
 import ai.h2o.sparkling.backend.utils.{ArgumentBuilder, RestApiUtils}
-import ai.h2o.sparkling.backend.{SharedBackendConf, SparklingBackend}
+import ai.h2o.sparkling.backend.{NodeDesc, SharedBackendConf, SparklingBackend}
 import ai.h2o.sparkling.utils.ScalaUtils._
 import ai.h2o.sparkling.utils.SparkSessionUtils
 import org.apache.commons.io.IOUtils
 import org.apache.spark.expose.Logging
-import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.h2o.{BuildInfo, H2OConf, H2OContext}
 import org.apache.spark.{SparkEnv, SparkFiles}
 import water.api.schemas3.CloudLockV3
