+++ core/src/main/scala/ai/h2o/sparkling/backend/utils/RestApiUtils.scala
@@ -19,8 +19,8 @@ package ai.h2o.sparkling.backend.utils
 
 import java.net.URI
 
+import ai.h2o.sparkling.backend.NodeDesc
 import org.apache.http.client.utils.URIBuilder
-import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.h2o.{H2OConf, H2OContext}
 import water.api.schemas3._
