+++ core/src/main/scala/ai/h2o/sparkling/backend/utils/SupportedTypes.scala
@@ -14,9 +14,9 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package org.apache.spark.h2o.utils
+package ai.h2o.sparkling.backend.utils
 
-import org.apache.spark.h2o.utils.ReflectionUtils.NameOfType
+import ai.h2o.sparkling.backend.utils.ReflectionUtils.NameOfType
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
 import water.fvec.Vec
@@ -25,9 +25,9 @@ import scala.language.{implicitConversions, postfixOps}
 import scala.reflect.runtime.universe._
 
 /**
-  * All type associations are gathered in this file.
-  */
-object SupportedTypes extends Enumeration {
+ * All type associations are gathered in this file.
+ */
+private[backend] object SupportedTypes extends Enumeration {
 
   type VecType = Byte
 
@@ -82,21 +82,18 @@ object SupportedTypes extends Enumeration {
 
   private def onNAreturn[T](value: T)(what: String) = value
 
-  // scalastyle:off
-  val Boolean   = SimpleType[scala.Boolean] (Vec.T_NUM,  BooleanType,   classOf[jl.Boolean  ], onNAreturn(false), typeOf[Boolean])
-  val Byte      = SimpleType[scala.Byte   ] (Vec.T_NUM,  ByteType,      classOf[jl.Byte     ], onNAthrow, typeOf[Byte])
-  val Short     = SimpleType[scala.Short  ] (Vec.T_NUM,  ShortType,     classOf[jl.Short    ], onNAthrow, typeOf[Short])
-  val Integer   = SimpleType[scala.Int    ] (Vec.T_NUM,  IntegerType,   classOf[jl.Integer  ], onNAthrow, typeOf[Int])
-  val Long      = SimpleType[scala.Long   ] (Vec.T_NUM,  LongType,      classOf[jl.Long     ], onNAthrow, typeOf[Long])
-  val Float     = SimpleType[scala.Float  ] (Vec.T_NUM,  FloatType,     classOf[jl.Float    ], onNAreturn(scala.Float.NaN), typeOf[Float])
-  val Double    = SimpleType[scala.Double ] (Vec.T_NUM,  DoubleType,    classOf[jl.Double   ], onNAreturn(scala.Double.NaN), typeOf[Double])
-  val Timestamp = SimpleType[js.Timestamp ] (Vec.T_TIME, TimestampType, classOf[js.Timestamp], onNAthrow)
-  val Date      = SimpleType[js.Date      ] (Vec.T_TIME, DateType,      classOf[js.Date     ], onNAthrow)
-  val String    = SimpleType[String       ] (Vec.T_STR,  StringType,    classOf[String],       onNAreturn(null), typeOf[String])
-  // TODO(vlad): figure it out, why
-  val UTF8      = SimpleType[UTF8String   ] (Vec.T_STR,  StringType,    classOf[String],       onNAreturn(null), typeOf[UTF8String])
-  // scalastyle:on
-  
+  val Boolean = SimpleType[scala.Boolean](Vec.T_NUM, BooleanType, classOf[jl.Boolean], onNAreturn(false), typeOf[Boolean])
+  val Byte = SimpleType[scala.Byte](Vec.T_NUM, ByteType, classOf[jl.Byte], onNAthrow, typeOf[Byte])
+  val Short = SimpleType[scala.Short](Vec.T_NUM, ShortType, classOf[jl.Short], onNAthrow, typeOf[Short])
+  val Integer = SimpleType[scala.Int](Vec.T_NUM, IntegerType, classOf[jl.Integer], onNAthrow, typeOf[Int])
+  val Long = SimpleType[scala.Long](Vec.T_NUM, LongType, classOf[jl.Long], onNAthrow, typeOf[Long])
+  val Float = SimpleType[scala.Float](Vec.T_NUM, FloatType, classOf[jl.Float], onNAreturn(scala.Float.NaN), typeOf[Float])
+  val Double = SimpleType[scala.Double](Vec.T_NUM, DoubleType, classOf[jl.Double], onNAreturn(scala.Double.NaN), typeOf[Double])
+  val Timestamp = SimpleType[js.Timestamp](Vec.T_TIME, TimestampType, classOf[js.Timestamp], onNAthrow)
+  val Date = SimpleType[js.Date](Vec.T_TIME, DateType, classOf[js.Date], onNAthrow)
+  val String = SimpleType[String](Vec.T_STR, StringType, classOf[String], onNAreturn(null), typeOf[String])
+  val UTF8 = SimpleType[UTF8String](Vec.T_STR, StringType, classOf[String], onNAreturn(null), typeOf[UTF8String])
+
   private implicit def val2type(v: Value): SimpleType[_] = v.asInstanceOf[SimpleType[_]]
 
   val allSimple: List[SimpleType[_]] = values.toList map val2type
