+++ core/src/main/scala/org/apache/spark/h2o/backends/internal/InternalBackendUtils.scala
@@ -18,9 +18,8 @@
 package org.apache.spark.h2o.backends.internal
 
 import ai.h2o.sparkling.backend.SharedBackendConf
-import ai.h2o.sparkling.backend.utils.{ArgumentBuilder, SharedBackendUtils}
+import ai.h2o.sparkling.backend.utils.{ArgumentBuilder, ReflectionUtils, SharedBackendUtils}
 import org.apache.spark.h2o.H2OConf
-import org.apache.spark.h2o.utils.ReflectionUtils
 import org.apache.spark.scheduler.local.LocalSchedulerBackend
 import org.apache.spark.{SparkContext, SparkEnv}
