+++ core/src/main/scala/org/apache/spark/h2o/backends/internal/InternalH2OBackend.scala
@@ -17,11 +17,10 @@
 
 package org.apache.spark.h2o.backends.internal
 
-import ai.h2o.sparkling.backend.SparklingBackend
+import ai.h2o.sparkling.backend.{NodeDesc, SparklingBackend}
 import ai.h2o.sparkling.backend.api.RestAPIManager
 import ai.h2o.sparkling.backend.external.ExternalBackendConf
 import ai.h2o.sparkling.utils.SparkSessionUtils
-import org.apache.spark.h2o.utils.NodeDesc
 import org.apache.spark.h2o.{H2OConf, H2OContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.rpc.RpcEndpointRef
